* refresh publisher connection on `Bunny::ChannelError` and `Timeout::Error` * remove deprecated `UnreadCountable` implementation * add `Stateful` coverage for all feed types * remove redundant tests for `Capped` feed * increment version number
