+++ lib/realself/feed.rb
@@ -9,7 +9,6 @@ require 'realself/feed/getable'
 require 'realself/feed/permanent'
 require 'realself/feed/redactable'
 require 'realself/feed/ttl'
-require 'realself/feed/unread_countable'
 require 'realself/feed/stateful'
 require 'realself/feed/state/bookmarkable'
 require 'realself/feed/state/sessioned'
