+++ lib/stream/version.rb
@@ -1,8 +1,8 @@
 module RealSelf
   module Stream
     MAJOR         = 4
-    MINOR         = 5
-    BUILD_NUMBER  = 3
+    MINOR         = 6
+    BUILD_NUMBER  = 0
 
     VERSION = "#{MAJOR}.#{MINOR}.#{BUILD_NUMBER}"
   end
